input: touchscreen: ft5435_mi: Fix the suspend/resume logic
The current touch panel suspend logic is very broken when gestures are enabled.

1. We should always use enable_irq_wake() in gesture mode to be sure that the IRQ
will work as a wakeup source. If we don't, there's a high chance that the gesture
will not be able to wake our system from a suspended state.

2. In the current state there's no separate resume logic for gesture mode at all.
This means that the driver will always try to wake the panel up from a sleep, which
makes no sense when wakeup gestures are enabled, as they need different registers for
the low-power mode.

Let's resolve the above problems, and while we're at it, also do some minor cleanups.

Change-Id: Icd402b9059b75aac4ce06ceb98bac80c1aa9c417

Partially revert "input: ft5435: Fix the suspend/resume logic"

* Apparently not every panel happy with that
* Let's revert most of the changes, but keep the IRQ-related fixes

This reverts commit 19f89492da6f2288b3a8a36d1b6fc89d66c87abc.

Change-Id: I7b4540075f23523037047b97e5fb05706f421bd7